migrated to use published sdk-react-native
