+++ app/settings.tsx
@@ -6,11 +6,9 @@ import { EmailInput } from "~/components/auth/auth.email";
 import { LoaderButton } from "~/components/ui/loader-button";
 import { PhoneInput } from "~/components/auth/auth.phone";
 import { useTurnkey } from '@turnkey/sdk-react-native';
-import { useAuthRelay } from '~/hooks/use-turnkey';
 
 const Settings = () => {
-  const { user } = useTurnkey();
-  const { updateUser } = useAuthRelay();
+  const { user, updateUser } = useTurnkey();
 
   const [email, setEmail] = useState<string>(user?.email || "");
   const [phone, setPhone] = useState<string>(user?.phoneNumber || "");
@@ -25,8 +23,15 @@ const Settings = () => {
   const { country, nationalNumber } = parsePhoneNumber(phone);
 
   const handleUpdateUser = async () => {
-    await updateUser({ email, phone });
-    Keyboard.dismiss();
+    try {
+      await updateUser({ email, phone });
+      Keyboard.dismiss();
+    }
+    catch (err){
+      alert("Failed to update user.");
+      console.error(err);
+    }
+   
   };
 
   const isDisabled =
