+++ components/account.tsx
@@ -1,22 +1,26 @@
-import * as React from 'react';
-import { Button } from '~/components/ui/button';
+import * as React from "react";
+import { Button } from "~/components/ui/button";
 import {
   DropdownMenu,
   DropdownMenuContent,
   DropdownMenuItem,
   DropdownMenuSeparator,
   DropdownMenuTrigger,
-} from '~/components/ui/dropdown-menu';
-import { Text } from '~/components/ui/text';
-import { Avatar } from './ui/avatar';
-import { useRouter } from 'expo-router';
-import { Drawer } from 'react-native-drawer-layout';
-import { useAuthRelay } from '~/hooks/use-turnkey';
-import DefaultProfile from '~/assets/svgs/default-profile.svg';
+} from "~/components/ui/dropdown-menu";
+import { Text } from "~/components/ui/text";
+import { Avatar } from "./ui/avatar";
+import { useRouter } from "expo-router";
+import { Drawer } from "react-native-drawer-layout";
+import { useTurnkey } from "@turnkey/sdk-react-native";
 
 export function Account() {
-  const { logout } = useAuthRelay();
+  const { clearSession } = useTurnkey();
   const router = useRouter();
+
+  const handleLogout = async () => {
+    await clearSession();
+  };
+
   return (
     <DropdownMenu>
       <DropdownMenuTrigger asChild>
@@ -34,12 +38,12 @@ export function Account() {
         sideOffset={12}
         className="w-64 native:w-72"
       >
-        <DropdownMenuItem onPress={() => router.push('/settings')}>
+        <DropdownMenuItem onPress={() => router.push("/settings")}>
           <Text>Settings</Text>
         </DropdownMenuItem>
 
         <DropdownMenuSeparator />
-        <DropdownMenuItem onPress={logout}>
+        <DropdownMenuItem onPress={handleLogout}>
           <Text>Log out</Text>
         </DropdownMenuItem>
       </DropdownMenuContent>
